Renamed LP.js to index.js
