+++ web/pages/index.js
@@ -1,12 +1,49 @@
-import MetrixCard from '../components/MetrixCard'
-import Button from '../components/Button'
+import Button from "../components/Button";
+import MetrixCard from "../components/MetrixCard";
+import styles from "../styles/LP.module.css";   
+import Link from 'next/link';
 
-export default function Home() {
-  return (
-    <div>
-      {/* <MetrixCard data={number:"120", sizeUnit:"mil", text:"Bmdia bom dia Bmdia bom dia Bmdia bom dia Bmdia bom dia"}></MetrixCard> */}
-
-      <Button isOutline={true} text="Saiba mais "></Button>
-    </div>
-  )
+export default function LP(){
+    return (
+        <div className="container">
+            <div className={styles.headerAndHero}>
+                <header>
+                    <img src="/images/lpLogo.svg" alt="beeco"/>
+                </header>
+                <div className={styles.hero}>
+                    <div className={styles.heroContent}>
+                        <h1 className={styles.heroTitle}>
+                            Entre para nossa colmeia
+                        </h1>
+                        <p className={styles.heroText}>
+                            Bee.co é a plataforma para os trabalhadores do futuro, entre na nossa colmeia e solucione os problemas mais abelhudos de milhões de usuarios. Se você precisa de uma abelha para te ajudar baixe agora e cadastre sua oferta de trabalho.
+                        </p>
+                        <Link href="/download">
+                            <Button isOutline={false} text="Baixe agora"></Button>
+                        </Link>
+                    </div>
+                    <img src="/images/lpHeroImg.svg" alt="Big bee" className={styles.heroImg}></img>
+                </div>
+            </div>
+            <div className={styles.metrixSection}>
+                <div>
+                    <MetrixCard data={{number: "150", 
+                                    sizeUnit: "Milhões", 
+                                    text: "Problemas solucionados por nossas abelhas"}}
+                                    extraClass ={styles.metrixCardLP}>
+                    </MetrixCard>
+                    <MetrixCard data={{number: "140", 
+                                    sizeUnit: "Mil", 
+                                    text: "De abelhas enriquecendo a cada dia"}}
+                                    extraClass = {styles.metrixCardLP}>
+                    </MetrixCard>
+                    <MetrixCard data={{number: "300", 
+                                    sizeUnit: "Reais", 
+                                    text: "Em movimento a cada minuto em nossa colmeia"}}
+                                    extraClass = {styles.metrixCardLP}>
+                    </MetrixCard>
+                </div>
+            </div>
+        </div>
+    );
 }
